LPR/requirements.txt
fastapi==0.110.2
filelock==3.6.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
load-dotenv==0.1.0
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.0
paho-mqtt==2.1.0
platformdirs==2.5.1
pydantic==2.7.1
pydantic_core==2.18.2
PyGObject==3.42.1
PyJWT==2.3.0
pymongo==4.7.1
pyparsing==2.4.7
